fastapi/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tensorflow>=2.13.0
numpy>=1.24.0
starlette>=0.27.0
python-multipart>=0.0.6
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
mypy>=1.4.0
